Refactoring package versioning
A new versioning scheme was needed for development (staging) versions. Previously these required guessing the next release and appending the GitLab CI pipeline ID. The former required manual updates as part of a release procedure.

Both this next version and ID are unstable, as the next version may switch between major, minor and patch depending on the issues that arise during development. The pipeline ID is specific to our instances of GitLab so is not safe to use across other platforms.

Having looked at other packages and the requirements for PyPi this project project now uses a deterministic and entirely automated version for both final and development releases.

This can be summarised as '[latest git tag].post[number of commits since tag].dev0' For example: '0.3.0.post12.dev0'. A script ran by CI is used to generate this version, which is saved to a text file and included in the source package. It is read by `setup.py` and `config.py` has been updated to read from the package metadata in production, or to use dummy values otherwise.

As this is new versioning scheme resolves around Git tags, Git is now available in the app container, and a version 0.1.0 tag has been made based on the commit that was incorrectly pushed to test.pypi as 0.1.0. This means 0.1.0 won't be available from PyPi.

Some other changes have been made to fix other settings and options, or to improve consistency.