Adding XML validation for WMS
Using a combination of the LXML Python library and XMLint command line utility for validating GetCapability documents against the WMS schema. WMS endpoints are validated automatically when the `flask data fetch` command is ran, or endpoints can be validated manually using a new command (`flask data validate`).

An XML catalog is used to cache schemas locally for quicker and more reliable tests.

Currently this requires running the container with root permissions so that GitLab can copy schema files from the build directory into the locations given in the catalog file. This is considered a bug and is tracked in #14.

Other changes include:

* refactoring building WMS endpoints from data source properties into a common function
* adding a JSON schema to validate the structure of data source files (i.e. that each source has a label)