Moving JSON and XML schemas inside package
To ensure data files are copied to right place when installed as a package.

This change splits the top-level ‘resources’ directory into a top-level ‘data’ directory, that holds runtime configuration/data,  and a static resources directory inside the package.

This requires the use of the importlib.resources module added in Python 3.7.