Fixing bandit nosec command
This was reformatted by Black and caused bandit to fail as the comment was in the wrong place (see https://github.com/PyCQA/bandit/issues/136).

Moving the #nosec comment to the first line rather than the end restores the override in Bandit and passes Black’s check so should be fine.