Fixing OGC interactive protocol selection in data validate CLI command
The choices list passed to Inquirer for prompting the user for a OGC protocol as part of the `data validate` CLI method was a string representation of the protocol enumeration, however the logic expected an actual enum element so caused a type mis-match.

This wasn’t caught by our tests because we patched the interactive inquirer call to return a static value, which was a valid enum element.

Now the choices list uses the value string of each enum element, and the patch method returns the same string. This value is used to lookup the relevant enum element after being selected (or patched).

Fixes #61.