Fixing aborting an inquirer prompt
Where an inquirer promot is aborted (the user presses [ctrl + c]) the answers element returned by the prompter is None rather than a dict of null answers.

This causes an exception when we check for a key in the prompter results dict (as it isn’t a dict). We now test for this and abort gracefully.

Fixes #60