Fixing XML catalogues
The deployment image is dependent on the current Python version so should have been changed when we upgraded to 3.8.

The name of the catalogue used in the development image was also inconsistent (through the proper spelling).

Adding README section for upgrading python versions so this doesn’t get missed in future.